Fix strategy save issues
The AI implemented the plan to fix strategy save issues. This involved correcting the `saveStrategy` function call in `EnhancedStrategyWizard.tsx` to use `fundId` instead of `fundType`, and resolving BigInt serialization errors by converting BigInt values to numbers before sending them to Supabase. The AI also verified the `enhancedCriteriaData` structure and tested the complete save flow to ensure data is correctly stored and displayed.

+++ src/services/strategyServiceV2.ts
@@ -86,9 +86,9 @@ class StrategyServiceV2 {
         
         // PHASE 1: Data type conversion and validation
         const checkSizeMin = wizardData.checkSizeRange?.min ? 
-          BigInt(Math.round(wizardData.checkSizeRange.min)) : null;
+          Math.round(wizardData.checkSizeRange.min) : null;
         const checkSizeMax = wizardData.checkSizeRange?.max ? 
-          BigInt(Math.round(wizardData.checkSizeRange.max)) : null;
+          Math.round(wizardData.checkSizeRange.max) : null;
           
         // PHASE 1: Convert fund type for V2 compatibility (use short form for V2 table)
         const convertFundType = (type: 'vc' | 'pe' | 'venture_capital' | 'private_equity'): 'vc' | 'pe' => {
